Added switch to use NtfsJunctions instead of copying for debug

+++ src/SamplePlugin/RegisterPluginAndStartStreamDeck.ps1
@@ -1,4 +1,11 @@
-﻿Write-Host "Gathering deployment items..."
+﻿# create a switch for the option to create and check a NTFS Junciton
+# Further named param must be declared here. This also leverages tap-completion on commandline.
+param (
+		[switch]$NtfsJunction = $false, # Adds the -NtfsJunction switch
+		[switch]$J = $NtfsJunction # Adds the -J switch (alias for -NtfsJunction)
+)
+
+Write-Host "Gathering deployment items..."
 
 Write-Host "Script root: $PSScriptRoot`n"
 
@@ -7,7 +14,6 @@ $basePath = $PSScriptRoot
 if (!($PSSCriptRoot)) {
   $basePath = $PWD.Path;
 }
-
 
 # Load and parse the plugin project file
 $pluginProjectFile = Join-Path $basePath "SamplePlugin.csproj"
@@ -21,11 +27,14 @@ $targetFrameworkName = $projectXML.Project.PropertyGroup.TargetFramework;
 
 # Set local path references
 if ($IsMacOS) {
+  $useJunction = $false # There is no NtfsJunction available on MacOS so we can NOT use it here
   $streamDeckExePath = "/Applications/Stream Deck.app"
   $bindir = "$basePath/bin/Debug/$targetFrameworkName/osx-x64"
-} else {
+}
+else {
   $streamDeckExePath = "$($ENV:ProgramFiles)\Elgato\StreamDeck\StreamDeck.exe"
   $bindir = "$basePath\bin\Debug\$targetFrameworkName\win-x64"
+  $useJunction = $j.IsPresent # Uses NtfsJunction when the -NtfsJunction switch is present
 }
 
 # Make sure we actually have a directory/build to deploy
@@ -42,9 +51,10 @@ $uuidAction = $json.Actions[0].UUID
 
 $pluginID = $uuidAction.substring(0, $uuidAction.Length - ".action".Length)
 
-if($IsMacOS) {
+if ($IsMacOS) {
   $destDir = "$HOME/Library/Application Support/com.elgato.StreamDeck/Plugins/$pluginID.sdPlugin"
-} else {
+}
+else {
   $destDir = "$($env:APPDATA)\Elgato\StreamDeck\Plugins\$pluginID.sdPlugin"
 }
 
@@ -52,15 +62,37 @@ $pluginName = Split-Path $basePath -leaf
 
 Get-Process -Name ("StreamDeck", $pluginName) -ErrorAction SilentlyContinue | Stop-Process –force -ErrorAction SilentlyContinue
 
-# Delete the target directory, make sure the deployment/copy is clean
-Remove-Item -Recurse -Force -Path $destDir -ErrorAction SilentlyContinue
-$bindir =  Join-Path $bindir "*"
+If (!$useJunction) {
+  # default behavior
 
-# Then copy all deployment items to the plugin directory
-New-Item -Type Directory -Path $destDir -ErrorAction SilentlyContinue # | Out-Null
-$bindir = $bindir +"\*"
-Copy-Item -Path $bindir -Destination $destDir -Recurse
+  # Delete the target directory, make sure the deployment/copy is clean
+  Remove-Item -Recurse -Force -Path $destDir -ErrorAction SilentlyContinue
 
+  $bindir = Join-Path $bindir "*"
+
+  # Then copy all deployment items to the plugin directory
+  New-Item -Type Directory -Path $destDir -ErrorAction SilentlyContinue # | Out-Null
+  $bindir = $bindir + "\*"
+  Copy-Item -Path $bindir -Destination $destDir -Recurse
+}
+else {
+  # checks if NtfsJunction exists and creates it, if it does not exist
+  if (!(Test-Path $destDir)) {
+    # $destDir doesn't exist
+    cmd /c mklink /j $destDir $bindir # create the junction pointing to $binddir
+    # using the cmd /c mklink because a "New-Item" needs admin rights!
+  }
+  else {
+    # something exists as $destDir
+    $junction = Get-Item $destDir
+    # Checks if the $destDir is a junction and is pointing to the correct $bindir
+    if (!($junction.Target.Contains($bindir))) {
+      # $destDir is either already a Junction but not targeting $bindir or it is someting else we can't use
+      Remove-Item -Recurse -Force -Path $destDir -ErrorAction SilentlyContinue # remove whatever $destDir is
+      cmd /c mklink /j $destDir $bindir # creates the junction to $binddir
+    }
+  }
+}
 
 Write-Host "Deployment complete. We will NOT restart the Stream Deck desktop application here, but will from the template..."
 # Start-Process $streamDeckExePath
